Fix button selection styling
Using .childNodes to select the spans on solidButton and randomButton
was breaking the previous implementation since the formatter inserted
whitespace into the HTML inside the button divs. This caused the
whitespace to be interpreted as text nodes, which cannot accept classes.
Switching to .children only returns the spans and ignores the
whitespace.